% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSRS.R
\name{NSRS}
\alias{NSRS}
\title{Check the native status for plant species in a political region}
\usage{
NSRS(occurrence_dataframe)
}
\arguments{
\item{occurrence_dataframe}{A properly formatted dataframe, see http://bien.nceas.ucsb.edu/bien/tools/nsr/batch-mode/}
}
\value{
Dataframe containing NSR results.
}
\description{
NSRS returns information on native status for species within a political region.
}
\examples{
\dontrun{
nsrs_testfile <- 
read.csv("http://bien.nceas.ucsb.edu/bien/wp-content/uploads/2019/02/nsr_testfile.csv")

results <- NSRS(occurrence_dataframe = nsrs_testfile)
  
# Inspect the results
head(results, 10)
# That's a lot of columns. Let's display one row vertically
# to get a better understanding of the output fields
results.t <- t(results[,2:ncol(results)]) 
results.t[,1,drop =FALSE]
# Summarize the main results
results[ 1:10, 
c("species", "country", "state_province", "native_status", "native_status_reason")]

# Compare summary flag isIntroduced to more detailed native_status values
# and inspect souces consulted
results[ 1:10, 
c("species", "country", "state_province", "native_status", "isIntroduced", "native_status_sources")]




}
}
