% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSRS_template.R
\name{NSRS_template}
\alias{NSRS_template}
\title{Make a template for an NSRS query}
\usage{
NSRS_template(nrow = 1)
}
\arguments{
\item{nrow}{The number of rows to include in the template}
}
\value{
Template data.frame that can be populated and then used in NSRS queries.
}
\description{
NSRS_template builds a template that can be populated to submit an NSRS query.
}
\examples{
\dontrun{

template<-NSRS_template(nrow = 2)
template$genus<-"Acer"
template$species<-c("Acer rubrum", "Acer saccharum")
template$country<-"Canada"
template$user_id<-1:2
results <- NSRS(occurrence_dataframe = template)

}
}
